Make Vulkan optional for encode libraries
This is a first pass at making Vulkan encode support optional by adding
a GFXRECON_ENABLE_VULKAN cmake option.
